FIX: Allow filtering to work with auto_select
Background:
https://github.com/seawolf/tty-prompt/blob/master/lib/tty/prompt/list.rb#L223-L225

The `List#choices` method only returns a `Choices` object if the list is original and unfiltered (ie. all choices are enabled) - otherwise it returns an array of choices.  So the safest way to determine that there is only one choice available is to check the size property on the return value directly.